[ Arch ] move files
