+++ app/src/main/java/architecture_components/utils/Objects.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package me.li2.android.wipro_assessment.utils;
+package architecture_components.utils;
 
-public class Objects {
+class Objects {
     public static boolean equals(Object o1, Object o2) {
         if (o1 == null) {
             return o2 == null;
