+++ app/src/main/java/architecture_components/utils/Status.java
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package me.li2.android.wipro_assessment.data.network;
+package architecture_components.utils;
 
 /**
  * Status of a resource that is provided to the UI.
  * <p>
  * These are usually created by the Repository classes where they return
  * {@code LiveData<Resource<T>>} to pass back the latest data to the UI with its fetch status.
+ *
+ * https://github.com/googlesamples/android-architecture-components/blob/master/GithubBrowserSample/app/src/main/java/com/android/example/github/vo/Status.java
  */
 public enum Status {
     SUCCESS,
