+++ app/src/main/java/me/li2/android/wipro_assessment/ui/countryintrolist/CountryIntroListFragmentViewModel.java
@@ -5,9 +5,9 @@ import android.arch.lifecycle.ViewModel;
 
 import java.util.List;
 
-import me.li2.android.wipro_assessment.data.network.Resource;
+import architecture_components.utils.Resource;
 import me.li2.android.wipro_assessment.data.repository.WiproRepository;
-import me.li2.android.wipro_assessment.data.database.CountryIntroEntry;
+import me.li2.android.wipro_assessment.data.model.CountryIntroEntry;
 
 /**
  * Created by weiyi on 04/02/2018.
